Linux installer (deb/run): only do kill -USR1 if the vboxweb-service init script is there because only VBoxSVC of these packages will catch SIGUSR1 properly
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@32542 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Installer/linux/routines.sh
@@ -97,7 +97,9 @@ check_root() {
 check_running() {
     VBOXSVC_PID=`pidof VBoxSVC 2> /dev/null`
     if [ -n "$VBOXSVC_PID" ]; then
-        kill -USR1 $VBOXSVC_PID
+        if [ -f /etc/init.d/vboxweb-service ]; then
+            kill -USR1 $VBOXSVC_PID
+        fi
         sleep 1
         if pidof VBoxSVC > /dev/null 2>&1; then
             echo 1>&2 "A copy of VirtualBox is currently running.  Please close it and try again."
